refactor: move AxisConstructor into axis mod
