+++ src/namespaces/builtins/mod.rs
@@ -1,4 +1,3 @@
-mod axis;
 mod circle;
 mod cube;
 mod cuboid;
@@ -8,7 +7,6 @@ mod plane;
 mod rectangle;
 mod sphere;
 
-pub use axis::AxisConstructor;
 pub use cube::Cube;
 pub use path::PathConstructor;
 pub use plane::PlaneConstructor;
