+++ src/values/mod.rs
@@ -1,12 +1,12 @@
-mod angle;
-mod axis;
-mod length;
-mod number;
-mod part;
-mod path;
-mod plane;
-mod sketch;
+pub mod angle;
+pub mod axis;
+pub mod length;
+pub mod number;
+pub mod part;
+pub mod path;
+pub mod plane;
+pub mod sketch;
 pub mod traits;
-mod value;
+pub mod value;
 
 pub use value::Value;
